More tests, splitting metadata test file.
